+++ src/simulator.rs
@@ -8,7 +8,7 @@ use petgraph::{
     dot::{Config, Dot},
     Graph,
 };
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::{fs::File, io::prelude::*, path::PathBuf};
 
 pub struct IdComponent(pub HashMap<String, Box<dyn Component>>);
@@ -32,12 +32,19 @@ impl Simulator {
 
         let mut id_nr_outputs = HashMap::new();
         let mut id_field_index = HashMap::new();
+
+        let mut sinks = vec![];
         // allocate storage for lensed outputs
 
         trace!("-- allocate storage for lensed outputs");
         for c in &component_store.store {
             trace!("{:?}", c.get_id_ports().0);
             let (id, ports) = c.get_id_ports();
+
+            // push all sinks
+            if c.is_sink() {
+                sinks.push(id.clone());
+            }
 
             trace!("id {}, ports {:?}", id, ports);
             // start index for outputs related to component
@@ -64,6 +71,8 @@ impl Simulator {
             }
             id_nr_outputs.insert(id.clone(), ports.outputs.len());
         }
+
+        trace!("sinks {:?}", sinks);
 
         let mut graph = Graph::<_, (), petgraph::Directed>::new();
         let mut id_node = HashMap::new();
@@ -161,6 +170,10 @@ impl Simulator {
             halt_on_warning: false,
             running_state: RunningState::Stopped,
             component_condition: vec![],
+            // used for determine active components
+            sinks,
+            inputs_read: HashMap::new(),
+            active: HashSet::new(),
         };
 
         trace!("sim_state {:?}", simulator.sim_state);
@@ -202,6 +215,28 @@ impl Simulator {
 
     /// get input value
     pub fn get_input_value(&self, input: &Input) -> SignalValue {
+        // trace!("get_input_value, input {:?}", input);
+
+        self.get_input_signal(input).get_value()
+    }
+
+    /// get input value and update set of inputs read
+    /// id, represents the component reading
+    /// input, represents the input it is reading
+    pub fn get_input_value_mut(&mut self, id: Id, input: &Input) -> SignalValue {
+        trace!("get_input_value_mut {:?} reading {:?}", id, input);
+
+        self.inputs_read
+            .entry(id)
+            .and_modify(|hs| {
+                hs.insert(input.id.clone());
+            })
+            .or_insert({
+                let mut hs = HashSet::new();
+                hs.insert(input.id.clone());
+                hs
+            });
+
         self.get_input_signal(input).get_value()
     }
 
@@ -254,13 +289,17 @@ impl Simulator {
             return;
         }
         // push current state
-        self.history.push(self.sim_state.clone());
+        self.history
+            .push((self.sim_state.clone(), self.active.clone()));
         trace!("cycle:{}", self.cycle);
+
+        self.clean_active();
 
         // clear component condition data for this new cycle
         self.component_condition.clear();
+
         for component in self.ordered_components.clone() {
-            //trace!("evaling component:{}", component.get_id_ports().0);
+            trace!("evaluating component:{}", component.get_id_ports().0);
 
             // Clock component and add its condition if error self.component_condition
             match component.clock(self) {
@@ -310,6 +349,44 @@ impl Simulator {
             }
         }
         self.cycle = self.history.len();
+        self.active_components()
+        // self.clock_mode = false;
+    }
+
+    // internal function to clear inputs read
+    fn clean_active(&mut self) {
+        trace!("clear_active");
+        self.inputs_read = HashMap::new();
+    }
+
+    // internal function to determine active components
+    fn active_components(&mut self) {
+        trace!("active - determine active components");
+        trace!("inputs read {:?}", self.inputs_read);
+
+        self.active = HashSet::new();
+
+        // iterate from sinks towards inputs
+        let mut to_visit = self.sinks.clone();
+
+        // extremely un-Rusty
+        while let Some(id) = to_visit.pop() {
+            if !self.active.contains(&id) {
+                trace!("id not found {}", id);
+                if let Some(ids) = self.inputs_read.get(&id) {
+                    trace!("reading input(s) {:?}", ids);
+                    for id in ids {
+                        to_visit.push(id.clone());
+                    }
+                }
+                self.active.insert(id);
+            }
+        }
+    }
+
+    /// check if component is active
+    pub fn is_active(&self, id: &Id) -> bool {
+        self.active.contains(id)
     }
 
     /// free running mode until Halt condition or target cycle, breaks after 1/30 sec
@@ -353,9 +430,11 @@ impl Simulator {
     /// reverse simulation using history if clock > 1
     pub fn un_clock(&mut self) {
         if self.cycle > 1 {
-            let state = self.history.pop().unwrap();
+            let (state, active) = self.history.pop().unwrap();
             // set old state
             self.sim_state = state;
+            self.active = active;
+
             // to ensure that history length and cycle count complies
             self.cycle = self.history.len();
             // TODO add component_condition history pop
@@ -368,15 +447,21 @@ impl Simulator {
 
     /// reset simulator
     pub fn reset(&mut self) {
+        // The order of the following is not important
+        // with the exception that self.clock() needs to be last
         self.history = vec![];
         self.cycle = 0;
-        self.sim_state.iter_mut().for_each(|val| *val = 0.into());
         self.running_state = RunningState::Stopped;
-        self.clock();
+        self.stop();
+
+        self.sim_state.iter_mut().for_each(|val| *val = 0.into());
+
         // TODO probably needed to reset component_condition, maybe is handeld correctly by clock who knows?
         for component in self.ordered_components.clone() {
             component.reset();
         }
+
+        self.clock();
     }
 
     // return the enum which describes the current state